requirements.txt
certifi==2024.2.2
deep-security-api @ https://automation.deepsecurity.trendmicro.com/sdk/20_0/v1/dsm-py-sdk.zip
python-dateutil==2.9.0.post0
six==1.16.0
urllib3==2.2.1
